+++ server/models/index.js
@@ -1,3 +1,5 @@
+'use strict';
+
 const fs = require("fs");
 const path = require("path");
 const Sequelize = require("sequelize");
